Skip unitMagnitude calculation when minValue < 2
In New(), unitMagnitude is set based on the input minValue. If the
input minValue is 0, math.Log2(0) returns -Inf, unitMagnitude is then
set by converting -Inf to an int32.

Right after, an IF statement checks whether unitMagnitude is less than
zero. If so, it is set to zero. This works fine in our x86
environment, however when running on MIPS, the IF statement returns
false, thus unitMagnitude is not reset, causing subsequent logic to
fail.

As the intent of the code is to set unitMagnitude to be at least zero,
do a check on the minValue first. If it is less than 2, simply leave unitMagnitude
as zero.